Move test id to own file
#402
