+++ trview.app.tests/trview.app.tests.vcxproj
@@ -58,6 +58,7 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
     </ClCompile>
     <ClCompile Include="TestImgui.cpp" />
+    <ClCompile Include="TestImGuiId.cpp" />
     <ClCompile Include="TriggersWindowManagerTests.cpp" />
     <ClCompile Include="TriggersWindowTests.cpp" />
     <ClCompile Include="UI\CameraControlsTests.cpp" />
@@ -100,6 +101,7 @@
     <ClInclude Include="NullImGuiBackend.h" />
     <ClInclude Include="stdafx.h" />
     <ClInclude Include="TestImgui.h" />
+    <ClInclude Include="TestImGuiId.h" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
     <VCProjectVersion>15.0</VCProjectVersion>
